ensure_nearby moved back to common in new module. Tests moved back to graph.rs.
